backend/services/monitoring/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-jose[cryptography]
django
djangorestframework
channels
